Reorg and rename mget
